pawsnet_backend/pawsnet/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-use-email-as-username==1.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==10.4.0
psycopg==3.2.5
python-dotenv==1.0.1
qrcode==8.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
